refactor: consolidate API URL definitions into dedicated route files
