+++ src/app/api/chatroom/chatroom/[chatroomId]/route.ts
@@ -1,7 +1,7 @@
-import { NextResponse } from 'next/server'
-import { getChatroomInfoURL } from '~/api/chatroom/api_url'
-import apiHandler from '~/api/api'
 import { cookies } from 'next/headers'
+import { NextResponse } from 'next/server'
+import apiHandler from '~/api/api'
+import { getChatroomInfoURL } from '~/api/chatroom/route'
 
 export async function GET(
     request: Request,
