+++ src/app/api/chatroom/chatroom/route.ts
@@ -1,7 +1,7 @@
-import { NextResponse } from 'next/server'
-import { createChatroomURL } from '~/api/chatroom/api_url'
-import apiHandler from '~/api/api'
 import { cookies } from 'next/headers'
+import { NextResponse } from 'next/server'
+import apiHandler from '~/api/api'
+import { createChatroomURL } from '~/api/chatroom/route'
 
 export async function POST(req: Request) {
     const { data } = await req.json()
