+++ src/app/api/user/get-user-info/route.ts
@@ -1,7 +1,7 @@
-import { NextResponse } from 'next/server'
-import { userInfoURL } from '~/api/user/api_url'
-import apiHandler from '~/api/api'
 import { cookies } from 'next/headers'
+import { NextResponse } from 'next/server'
+import apiHandler from '~/api/api'
+import { userInfoURL } from '~/api/user/route'
 
 export async function GET() {
     const token = cookies().get('auth_token')?.value
